+++ core/src/main/java/io/snabble/sdk/PaymentMethod.java
@@ -9,6 +9,8 @@ public enum PaymentMethod {
     ENCODED_CODES(true, false),
     @SerializedName("encodedCodesCSV")
     ENCODED_CODES_CSV(true, false),
+    @SerializedName("encodedCodesIKEA")
+    ENCODED_CODES_IKEA(true, false),
     @SerializedName("deDirectDebit")
     DE_DIRECT_DEBIT(false, true);